backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
reportlab==4.2.5
Pillow==11.0.0
